requirements.txt
asgiref==3.11.0
black==25.11.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Django==5.2.8
django-redis==6.0.0
djangorestframework==3.16.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
pytokens==0.3.0
redis==7.1.0
sqlparse==0.5.3
